requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
countryinfo==0.1.2
Django==5.2.6
django-countries==7.6.1
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
phonenumbers==9.0.13
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==5.2.1
requests==2.32.5
responses==0.25.8
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
